iris-server/requirements.txt
Flask==0.10.1
Flask-Cors==2.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
numpy==1.9.2
pandas==0.16.2
Pillow==2.9.0
python-dateutil==2.4.2
pytz==2015.4
scikit-learn==0.16.1
scipy==0.16.0
six==1.9.0
Werkzeug==0.10.4
wheel==0.24.0
